change formatting (#250)

+++ cleanup-docker.sh
@@ -1,3 +1,21 @@
-docker image ls | grep odyssey | awk '{print $3}' | xargs docker image rm -f || true
-docker rm $(docker stop $(docker ps -aq)) || true
-docker network rm $(docker network ls | grep jepsen | egrep -o '^[^ ]+') || true
+#!/bin/bash
+
+set -ex 
+
+while getopts "h:s:" arg; do
+  case $arg in
+    h)
+      echo "usage" 
+      ;;
+    s)
+      skip=$OPTARG
+      echo $skip
+      ;;
+  esac
+done
+
+
+if [ -z $skip ]; then
+    docker image ls | grep odyssey | awk '{print $3}' | xargs docker image rm -f || true
+    docker rm $(docker stop $(docker ps -aq)) || true
+fi
